requirements.txt
streamlit
langchain
langchain-openai
requests
python-dotenv
google-search-results
numpy